requirements.txt
pandas
numpy
lightgbm
scikit-learn
pyodbc
python-dotenv
